Move upload form to its own component
